requirements.txt

##  محتوى requirements.txt
```txt
pandas
scikit-learn
faiss-cpu
sentence-transformers
streamlit
